cloud-functions/enc-converter/requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
google-cloud-storage>=2.14.0
google-cloud-firestore>=2.14.0
google-cloud-run>=0.10.0
requests>=2.31.0
google-auth>=2.27.0
aiohttp>=3.9.0
GDAL==3.4.1